Added status conversion
